+++ packages/plasma-core/src/components/Tabs/Tabs.component-test.tsx
@@ -1,7 +1,5 @@
 import React from 'react';
-import ReactDom from 'react-dom';
-import { mount } from '@cypress/react';
-import { CypressTestDecorator, getComponent } from '@sberdevices/plasma-cy-utils';
+import { mount, CypressTestDecorator, getComponent } from '@sberdevices/plasma-cy-utils';
 
 describe('plasma-core: Tabs', () => {
     const Tabs = getComponent('Tabs');
@@ -16,7 +14,6 @@ describe('plasma-core: Tabs', () => {
                     <TabItem>Eva</TabItem>
                 </Tabs>
             </CypressTestDecorator>,
-            { ReactDom },
         );
 
         cy.matchImageSnapshot();