+++ packages/plasma-ui/src/components/Badge/Badge.component-test.tsx
@@ -1,8 +1,6 @@
 import React from 'react';
-import ReactDom from 'react-dom';
-import { mount } from '@cypress/react';
 import { IconEye } from '@sberdevices/plasma-icons';
-import { CypressTestDecorator, PadMe, SpaceMe } from '@sberdevices/plasma-cy-utils';
+import { mount, CypressTestDecorator, PadMe, SpaceMe } from '@sberdevices/plasma-cy-utils';
 
 import { Badge } from '.';
 
@@ -22,7 +20,6 @@ describe('Badge', () => {
                     <Badge view="secondary" contentLeft={<Icon />} />
                 </div>
             </CypressTestDecorator>,
-            { ReactDom },
         );
         cy.matchImageSnapshot();
     });