+++ packages/plasma-ui/src/components/Button/Button.component-test.tsx
@@ -1,7 +1,6 @@
 import React from 'react';
 import ReactDom from 'react-dom';
-import { mount } from '@cypress/react';
-import { CypressTestDecorator } from '@sberdevices/plasma-cy-utils';
+import { mount, CypressTestDecorator } from '@sberdevices/plasma-cy-utils';
 
 import { Button } from './Button';
 
@@ -22,7 +21,6 @@ describe('Button', () => {
                     stretch={false}
                 />
             </CypressTestDecorator>,
-            { ReactDom },
         );
         cy.matchImageSnapshot();
     });