+++ packages/plasma-web/src/components/Progress/Progress.component-test.tsx
@@ -1,7 +1,5 @@
 import React from 'react';
-import ReactDom from 'react-dom';
-import { mount } from '@cypress/react';
-import { CypressTestDecorator, getComponent, PadMe } from '@sberdevices/plasma-cy-utils';
+import { mount, CypressTestDecorator, getComponent, PadMe } from '@sberdevices/plasma-cy-utils';
 
 describe('plasma-web: Progress', () => {
     const Progress = getComponent('Progress');
@@ -11,7 +9,6 @@ describe('plasma-web: Progress', () => {
             <CypressTestDecorator>
                 <Progress value={25} status="success" />
             </CypressTestDecorator>,
-            { ReactDom },
         );
 
         cy.matchImageSnapshot();
@@ -22,7 +19,6 @@ describe('plasma-web: Progress', () => {
             <CypressTestDecorator>
                 <Progress value={25} displayValue={false} status="success" />
             </CypressTestDecorator>,
-            { ReactDom },
         );
 
         cy.matchImageSnapshot();
@@ -37,7 +33,6 @@ describe('plasma-web: Progress', () => {
                 <PadMe />
                 <Progress value={100} status="error" />
             </CypressTestDecorator>,
-            { ReactDom },
         );
 
         cy.matchImageSnapshot();